fix(axiom): keep the allowlist off the client bundle, and declare vitest
Two CI failures, both introduced by this PR.

CLIENT GRAPH. `no-server-infra-in-app-graph` failed: the new `datastreams.ts`
was a NEW node on an EXISTING client-reachable chain —

  _app.tsx -> [dyn] feature-flags.service -> [dyn] flipt/client
    -> server/logging/client -> civitai-axiom/client -> civitai-axiom/env
    -> civitai-axiom/datastreams

`env.ts` was already on that graph with a reviewed exception; importing a new
module from it dragged the allowlist into every browser bundle. A dynamic
import() does not exempt a module — the chunk is still compiled.

Fixed by folding the constants back into `env.ts` and deleting the module, so
the client graph gains ZERO nodes. The alternative — an entry in the guard's
KNOWN_REACHABLE list — was rejected: that list is documented as one that "can
shrink but never silently grow", and growing it to accept a violation THIS
change introduced is the repair the guard exists to prevent. Narrowing
DENIED_PREFIXES the way @civitai/buzz already is would also work (these are pure
constants with no imports) but edits a shared safety mechanism; recorded as a
follow-up instead. A header comment in env.ts records why the split must not be
re-done.

ESLINT. `vitest` was not in the package's devDependencies. Every other package
with tests declares it; @civitai/axiom was the sole omission, invisible because
the gate lints ADDED files only — so it fired on this PR's two new test files
and would fire on anyone else's. Declared it, and regenerated the lockfile entry
so `--frozen-lockfile` still resolves. This also clears the same error on the two
pre-existing test files.

Verified: `no-server-infra-in-app-graph` 5/5 green; the seam mutations still kill
after the move (empty set -> 5 failed, extras dropped -> 2 failed, same
assertions as before); typecheck error set byte-identical to the origin/main
baseline; eslint clean across the whole package.
