+++ packages/civitai-axiom/package.json
@@ -7,5 +7,8 @@
   "dependencies": {
     "@axiomhq/axiom-node": "^0.12.0",
     "zod": "^4.0.17"
+  },
+  "devDependencies": {
+    "vitest": "^4.1.11"
   }
 }
