+++ packages/civitai-axiom/src/index.ts
@@ -1,2 +1,1 @@
 export * from './client';
-export * from './datastreams';
